pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["pedrohbsilva <phbs1235@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
Flask = "^2.1.3"
Flask-SQLAlchemy = "^2.5.1"
psycopg2 = "^2.9.3"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.28.1"
Flask-Migrate = "^3.1.0"
flasgger = "^0.9.5"
bcrypt = "^3.2.2"
requests = "^2.28.1"
PyJWT = "^2.4.0"
google-auth-oauthlib = "^0.5.2"
google-auth = "^2.10.0"
Flask-Cors = "^3.0.10"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-flask = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
